requirements.txt
beautifulsoup4==4.12.3
bitsandbytes==0.43.2
diffuser==0.0.1
-e git+https://github.com/huggingface/diffusers@bbd2f9d4e9ae70b04fedf65903fd1fb035437db4#egg=diffusers
gradio==4.38.1
gradio_client==1.1.0
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
